third_party/blink/web_tests/platform/mac-mac10.13/virtual/scroll-unification/plugins/focus-change-4-change-focus-and-blur-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(false)
i1focus
i1blur
pfocus-change-2
pblur-change-4
i4focus
i4blur
i3focus
PASS successfullyParsed is true

TEST COMPLETE


Test changing focus inside plugin onfocus and onblur events

    
